Update wscript
I think the if statement "if host in ['mingw32']" should be "if 'mingw32' in host" because, currently, if your toolchain name is i686-w64-mingw32, it does not match. Actually my rtems source builder fails because of this and it's been failing both on my LFS machine and my Ubuntu 16 server for that reason.